% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult tissues}
\format{An object of class \code{"SummarizedExperiment"}; see 
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
\usage{
data(adult)
}
\description{
A dataset containg expression measures for different adult tissues.
}
\examples{
data(adult)
adult
}
\keyword{datasets}
