% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fetal_wo}
\alias{fetal_wo}
\title{Fetal tissues}
\format{An object of class \code{"SummarizedExperiment"}; see 
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
\usage{
data(fetal_wo)
}
\description{
A dataset containg expression measures for different fetal tissues. Includes
samples from 1st and 2nd trimester.
}
\examples{
data(fetal_wo)
fetal_wo
}
\keyword{datasets}
