\name{KeyGenes-package}
\alias{KeyGenes-package}
\alias{KeyGenes}
\docType{package}
\title{
\packageTitle{KeyGenes}
}
\description{
\packageDescription{KeyGenes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KeyGenes}
\packageIndices{KeyGenes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{KeyGenes}

Maintainer: \packageMaintainer{KeyGenes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
