% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{keygenes.dendrogram}
\alias{keygenes.dendrogram}
\title{KeyGenes dendrogram}
\usage{
keygenes.dendrogram(data)
}
\arguments{
\item{data}{A KeyGenesResults containing prediction results.}
}
\value{
A ggplot displaying a dendrogram of the training and test data.
}
\description{
Draw a dendrogram for KeyGenes results.
}
\examples{
#' data(adult)
data(fetal_wo)
result <- keygenes.NGS(adult, fetal_wo, "tissue")
keygenes.dendrogram(result)
}
