% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{keygenes.heatmap}
\alias{keygenes.heatmap}
\title{Title}
\usage{
keygenes.heatmap(data, clusterTissues = F, clusterSamples = F)
}
\arguments{
\item{data}{blah}

\item{clusterTissues}{blah}

\item{clusterSamples}{blah}
}
\value{
blah
}
\description{
Title
}
\examples{
blah
}
