% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygenes.NGS.R
\name{keygenes.NGS.run}
\alias{keygenes.NGS.run}
\title{Title}
\usage{
keygenes.NGS.run(test, train, train.classes, genes, test.classes,
  verbose = FALSE)
}
\arguments{
\item{test}{blah}

\item{train}{blah}

\item{train.classes}{blah}

\item{genes}{blah}

\item{test.classes}{blah}

\item{verbose}{blah}
}
\value{
blah
}
\description{
Title
}
\examples{
blah
}
