% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygenes.R
\docType{class}
\name{KeyGenesResults-class}
\alias{KeyGenesResults-class}
\title{KeyGenes prediction results}
\description{
KeyGenes prediction results
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame. The predictions.}

\item{\code{accuracy}}{numeric.  Accuracy of the predictions or NaN.}

\item{\code{cvfit}}{ANY. The model used for the predictions.}

\item{\code{class.genes}}{list. The key genes for each class.}

\item{\code{prediction.matrix}}{matrix. The raw prediction values.}

\item{\code{train}}{matrix. The final effective training set.}

\item{\code{train.classes}}{character. The classes for each sample in the training set.}

\item{\code{test}}{matrix. The final effective test set.}

\item{\code{test.classes}}{OptionalChar.  The classes for the test set.}

\item{\code{genes}}{character. The genes used for training and prediction.}

\item{\code{duplicated.sample}}{integer. The indexes of the samples which which were
duplicated in order to have a enough samples per class for training the model.}
}}

