% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostVariableGenes.R
\name{mostVariableGenes}
\alias{mostVariableGenes}
\title{Get top n most variable genes}
\usage{
mostVariableGenes(counts, n = 500)
}
\arguments{
\item{counts}{A numeric matrix containing raw expression counts.}

\item{n}{The number of genes to be returned. Defaults to 500.}
}
\value{
A vector containing gene identifiers.
}
\description{
Retrieve the list of genes identifiers for the genes which have the most 
variable expression in the given counts (based on voom normalized counts).
Rownames of the given count matrix are assumed to be gene identifiers.
}
\examples{
top500 <- mostVariableGenes(counts)
}
