% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{training.data}
\alias{training.data}
\title{Training data}
\format{An object of class \code{"SummarizedExperiment"}; see 
\code{\link[SummarizedExperiment]{SummarizedExperiment}}.}
\usage{
data(training.data)
}
\description{
A dataset containing expression measures for different organs at different
stages of maturity. Containg hESC samples (Forster et al., 2014; Roost et 
al., 2015), 1st and 2nd trimester fetal samples (Roost et al., 2015) and 
adult samples (Cnop et al., 2014; Fagerberg et al., 2014; Illumina Body Map
2.0). Counts may not be available for all genes for all samples.
}
\examples{
data(training.data)
training.data

# Get only those genes for which each sample has a value.
training.data[complete.cases(assay(training.data)),]

}
\references{
Cnop, M., Abdulkarim, B., Bottu, G., Cunha, D.A., Igoillo-Esteve, M., Masini, M., Turatsinze, J.V., Griebel, T., Villate, O., Santin, I., et al. (2014). RNA sequencing identifies dysregulation of the human pancreatic islet transcriptome by the saturated fatty acid palmitate. \emph{Diabetes}. 63(6): 1978-1993.

Fagerberg, L., Hallstrom, B.M., Oksvold, P., Kampf, C., Djureinovic, D., Odeberg, J., Habuka, M., Tahmasebpoor, S., Danielsson, A., Edlund, K., et al. (2014). Analysis of the human tissue-specific expression by genome-wide integration of transcriptomics and antibody-based proteomics. \emph{Molecular & cellular proteomics: MCP}. 13(2): 397-406.

Forster R., Chiba K., Shaeffer L., Regalado S.G., Lai C.S., Gao Q., Kiani S., Farin H.F., Clevers H., Cost G.J., et al. (2014) Human intestinal tissue with adult stem cell properties derived from pluripotent stem cells. \emph{Stem cell reports}. 2(6): 838-852.

Roost M.S., van Iperen L., Ariyurek Y., Buermans H.P., Arindrarto W., Devalla H.D., Passier R., Mummery C.L., Carlotti F., de Koning E.J.P., van Zwet E.W., Goeman J.J., and Chuva de Sousa Lopes S.M. (2015). KeyGenes, a tool to probe tissue differentiation using a human fetal transcriptional atlas. \emph{Stem Cell Reports}. 4(6):1112-24.
}
\keyword{datasets}
