\name{KeyGenes-package}
\alias{KeyGenes-package}
\alias{KeyGenes}
\docType{package}
\title{
\packageTitle{KeyGenes}
}
\description{
\packageDescription{KeyGenes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KeyGenes}
\packageIndices{KeyGenes}
}
\author{
Maintainer: \packageMaintainer{KeyGenes}
}
\references{
Friedman J., Hastie T., and Tibshirani R. (2010). Regularization paths for 
generalized linear models via coordinate descent. \emph{Journal of statistical 
software}. 33(1): 1-22.

Roost M.S., van Iperen L., Ariyurek Y., Buermans H.P., Arindrarto W., 
Devalla H.D., Passier R., Mummery C.L., Carlotti F., de Koning E.J.P., 
van Zwet E.W., Goeman J.J., and Chuva de Sousa Lopes S.M. (2015). KeyGenes, 
a tool to probe tissue differentiation using a human fetal transcriptional
atlas. \emph{Stem Cell Reports}. 4(6):1112-24.
}
\keyword{ package }
\seealso{
\code{\link[glmnet:glmnet-package]{glmnet}}
}
