% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{keygenes.heatmap}
\alias{keygenes.heatmap}
\title{Keygenes heatmap}
\usage{
keygenes.heatmap(data, clusterClasses = F, clusterSamples = F)
}
\arguments{
\item{data}{A KeyGenesResults containing prediction results.}

\item{clusterClasses}{Whether or not to cluster the classes. Defaults to FALSE.}

\item{clusterSamples}{Whether or not to cluster the samples. Defaults to FALSE.}
}
\value{
A ggplot displaying a heatmap of the predictions.
}
\description{
Draw a heatmap for KeyGenes results.
}
\examples{
data(adult)
data(fetal_wo)
result <- keygenes.NGS(adult, fetal_wo, "tissue")
keygenes.heatmap(result)
}
