% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygenes.NGS.R
\name{keygenes.NGS}
\alias{keygenes.NGS}
\title{Predict the class (eg. organ) for NGS (RNA-seq) samples}
\usage{
keygenes.NGS(
  test,
  train,
  train.classes,
  genes = NULL,
  test.classes = NULL,
  confidence.cutoff = 0.7,
  verbose = FALSE
)
}
\arguments{
\item{test}{The counts table of the test data or a SummarizedExperiment 
containing it.}

\item{train}{The counts table of the training data or a SummarizedExperiment
containing it.}

\item{train.classes}{The class each sample in the training set belongs to.
Each sample from the test set will be assigned to one of these. Normally a
character vector with one item per train sample, if `train` is a 
SummarizedExperiment the colData column containing the classes.}

\item{genes}{The genes to be used for training and testing. If NULL
\link[KeyGenes]{most.variable.genes} will be called wih n = 500. Defaults to
NULL}

\item{test.classes}{See train.classes, but for the test samples. Used to
calculate the accuracy of the predictions. If NULL the accuracy will be NaN.
Defaults to NULL}

\item{confidence.cutoff}{The minimum prediction value for a prediction to be
considered high confidence.}

\item{verbose}{Should progress be reported, defaults to FALSE}
}
\value{
A \link[KeyGenes]{KeyGenesResults} object containing the prediction 
results.
}
\description{
Predict the class (eg. organ or maturation) for the test samples based on
a set of training samples, using RNA-seq expression measures.
}
\examples{
data(adult)
data(fetal_wo)
result <- keygenes.NGS(adult, fetal_wo, "tissue")
}
