% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygenes.NGS.R
\name{keygenes.NGS.run}
\alias{keygenes.NGS.run}
\title{Predict the class (eg. organ) for NGS (RNA-seq) samples}
\usage{
keygenes.NGS.run(test, train, train.classes, genes, test.classes,
  verbose = FALSE)
}
\arguments{
\item{test}{A matrix of expression measures for the test set}

\item{train}{A matrix of expression measures for the training set}

\item{train.classes}{A vector with the classes of the training samples. Each
sample from the test set will be assigned to one of these}

\item{genes}{The genes to be used for training and testing}

\item{test.classes}{A vector with the classes of the test samples. Used to
calculate accuracy. May be NULL}

\item{verbose}{Should progress be reported, defaults to FALSE}
}
\value{
A  \link[KeyGenes]{KeyGenesResults} object containing the prediction 
results.
}
\description{
Predict the class (eg. organ or maturation) for the test samples based on
a set of training samples, using RNA-seq expression measures.
}
