% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keygenes.R
\docType{class}
\name{KeyGenesResults-class}
\alias{KeyGenesResults-class}
\title{Title}
\value{
blah
}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{result}}{data.frame.}

\item{\code{accuracy}}{numeric.}

\item{\code{cvfit}}{ANY.}

\item{\code{class.genes}}{list.}

\item{\code{prediction.matrix}}{matrix.}

\item{\code{train}}{matrix.}

\item{\code{train.classes}}{character.}

\item{\code{test}}{matrix.}

\item{\code{test.classes}}{OptionalChar.}

\item{\code{genes}}{character.}
}}

\examples{
blah
}
